update: friend_add_update

See merge request s11-metaverse-game-sub1/S11P21D110!32

+++ server/src/main/java/com/dropzone/friend/dto/WaitingFriendListDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WaitingFriendListDto {
-    private Long friendShipId;
+    private int friendShipId;
     private String friendEmail;
     private String friendNickname;
     private FriendShipStatus status;
