+++ server/src/main/java/com/dropzone/friend/repository/FriendShipRepository.java
@@ -3,6 +3,6 @@ package com.dropzone.friend.repository;
 import com.dropzone.friend.entity.FriendShipEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FriendShipRepository extends JpaRepository<FriendShipEntity, Long> {
+public interface FriendShipRepository extends JpaRepository<FriendShipEntity, Integer> {
 
 }
